chore: move latest gh release finder to utils
